+++ examples/evolution/xor/main.rs
@@ -14,8 +14,7 @@
  * ```
  */
 
-use only_torch::nn::evolution::gene::TaskMetric;
-use only_torch::nn::evolution::{Evolution, EvolutionResult};
+use only_torch::nn::evolution::{Evolution, EvolutionResult, TaskMetric};
 use only_torch::tensor::Tensor;
 use std::path::Path;